changed id datatype

+++ src/main/java/pl/foodoutlet/foodoutlet/repository/FoodOutletRepo.java
@@ -3,6 +3,6 @@ package pl.foodoutlet.foodoutlet.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pl.foodoutlet.foodoutlet.model.FoodOutlet;
 
-public interface FoodOutletRepo extends JpaRepository<FoodOutlet, Integer> {
+public interface FoodOutletRepo extends JpaRepository<FoodOutlet, Long> {
 
 }
